Changed rxTCP, txTCP, rxUDP, txUDP to use the new Socket Memory variable.  This is the final step in implementing a new routine for adjusting socket memory size in the W5100.
'***************************************
PUB InitSocketMem( _template) | temp0
'***************************************
'' Initialize the socket memory of the W5100.  The W5100 has a total of 8kb for rx memory and 8kb for tx memory.
'' This memory can be assigned in 1kb, 2kb, 4kb and 8kb increments.  Memory assignments starts at socket 0 and work
'' towards socket 3 and if there isn't enough memory left for a socket, the socket doesn't get any and shouldn't
'' be used.
''
'' Users are welcome to add their own templates to fit their design.
''
'' This design assumes that the rx and tx registers are identically sized, but they don't have to be.
''
''  params:  _template              
''
''  return:  none